Made changes based on comments left on PR #116
Changed user_id to userId in App.js, put getRolesFromClaims in
services/claims.js and made functions isAnOfficer(userContext),
isAMember(userContext) and isAnInductee(userContext) in
services/claims.js as well.
